backend/rag_backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
qdrant-client==1.8.0
python-dotenv==1.0.0
PyYAML==6.0.1
google-generativeai==0.8.5
tiktoken==0.5.2
python-multipart==0.0.6
pydantic==2.5.0
openai==1.3.5
asyncpg==0.30.0
anthropic==0.21.3